// src/app/enums/aircraft-types.ts
export enum AircraftTypes {
  UF10 = 'UFM-10 Samba',
  T204 = 'Tu-204',
  T154 = 'Tu-154',
  T134 = 'Tu-134',
  SW4 = 'C-26D Metro 23',
  R135 = 'RC-135 Rivet Joint',
  PZ4M = 'PZL 104M Wilga 2000',
  PZ3T = 'PZL-130 Orlik',
  PA18 = 'Piper L-21B Super Cub',
  P06T = 'P-2006T',
  MI8 = 'Mi-24',
  M339 = 'MB-339CD',
  LJ35 = 'C-21A',
  L410 = 'L-410 Turbolet',
  K35R = 'KC-135T Stratotanker',
  IL76 = 'Il-76',
  IL62 = 'Il-62',
  H64 = ' WAH-64D Apache AH1',
  H60 = 'Sikorsky UH-60M Black Hawk',
  H47 = 'CH-47F Chinook',
  GLF5 = 'GV-SP (G550)',
  GLF4 = 'S 102B Korpen',
  GLEX = 'Bombardier E-11A',
  F900 = 'Falcon 900EX',
  F15 = 'F-15 Eagle',
  EUFI = 'Typhoon FGR4',
  EN48 = 'Enstrom 480B-G',
  DRTG = '',
  E3CF = 'E-3A Sentry',
  E121 = 'EMB 121A Xingu',
  DC10 = 'KC-10A Extender',
  D228 = 'UC-228',
  CULV = 'Sikorsky UH-60M Black Hawk',
  CN35 = 'CASA CN-235M-200',
  CL60 = 'CL-600-2B16 Challenger 604',
  C560 = 'UC-35A Citation Ultra',
  C510 = '510 Citation Mustang',
  C30J = 'C-130J Hercules C4',
  C295 = '295M',
  C27J = 'C-27J Spartan',
  C17 = 'C-17 Globemaster III',
  C130 = 'C-130 Hercules',
  BE20 = 'RC-12X Huron',
  B763 = 'KC-46A Pegasus',
  B738 = '737-800WL',
  B737 = 'C-40C/737',
  B703 = 'E-8C J-Stars',
  B52 = 'B-52H Stratofortress',
  B429 = '429 GlobalRanger',
  B350 = 'UC-12W Huron/Shadow R1',
  AT45 = 'ATR 42-500MP Surveyor',
  AS65 = 'AS365 N2 Dauphin 2',
  AN32 = 'An-32',
  AN28 = 'M28B/PT Bryza',
  AN26 = 'An-26',
  AN12 = 'An-12BK',
  A400 = 'A400M-180',
  A332 = 'Voyager KC3',
  A148 = 'An-148',
  A124 = 'An-124-100 Ruslan',
}